Changed the data type of the loop counter variable to avoid overflow which results in property violation in execution based tools

+++ c/array-fpi/ss2.c
@@ -56,7 +56,7 @@ int main()
 
 	for(i=0; i<N; i++)
 	{
-		__VERIFIER_assert(a[i] == N*(N+2));
+		__VERIFIER_assert(a[i] == (long long)N*(N+2));
 	}
 	return 1;
 }
